Updated rockin to support puma application server
Added capistrano scripts to support installation and deployment on to
Ubuntu 12.04+. Puma is started via Upstart. Information on the
configuration of Upstart can be found here:
https://github.com/puma/puma/tree/master/tools/jungle/upstart

The rails generator has been updated to allow for the selection of puma,
trinidad or unicorn application servers.

Separate configuration file templates have been created for puma,
trinidad and unicorn configuration of nginx.