Changes to base, postgres, rbenv and unicorn recipies
Updated base to include the option timezone. This option is needed when
deploying to AWS, Amazon will not allow you to tweak the timezone on the
server.

rbenv.rb was updated to include a jdk_version for JRuby deployments.
This allows the user to choose between jdks such as openjdk-6-jdk or
openjdk-7-jdk.

postgres.rb has been updated to automatically install postgres-contrib
and create the hstore extsion. This as been added to the library because
migrations that are not run a postgres installing extensions will fail,
causing the user to comment out the migration, install the extension,
and deploy/migrate again.

unicorn.rb was updated to use sudo when restarting unicorn. This caused
applications in production to not alway restart and will ensure restarts
after deploys in the future.